# Speedup resource count calculation