@blueorangutan 2024-04-13
<b>[SF] Trillian test result (tid-9811)</b>
Environment: kvm-centos7 (x2), Advanced Networking with Mgmt server 7
Total time taken: 48680 seconds
Marvin logs: https://github.com/blueorangutan/acs-prs/releases/download/trillian/pr8903-t9811-kvm-centos7.zip
Smoke tests completed. 129 look OK, 0 have errors, 0 did not run
Only failed and skipped tests results shown below:


Test | Result | Time (s) | Test File
--- | --- | --- | ---